Adicionando mais modelos ao projeto

+++ graduacao/src/main/java/br/usp/exemplo/graduacao/repositorios/MatriculaRepositorio.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface MatriculaRepositorio extends JpaRepository<Matricula, Integer> {
+public interface MatriculaRepositorio extends JpaRepository<Matricula, Long> {
 }
